refactor(device/live): 移除 WithApp后缀方法
appName可以在实例化Live的时候已经设置，
后面几乎不会更改，所以移除了WithApp后缀方法。
如果必须中途修改appName，可以通过Live.SetAppName()
更换。后续新接口也不再提供WithApp后缀方法。
