+++ device/live/live.go
@@ -86,6 +86,11 @@ const (
 	DescribeLiveStreamsNotifyUrlConfigAction = "DescribeLiveStreamsNotifyUrlConfig"
 	DeleteLiveStreamsNotifyUrlConfigAction   = "DeleteLiveStreamsNotifyUrlConfig"
 
+	// 直播转点播
+	AddLiveRecordVodConfigAction      = "AddLiveRecordVodConfig"
+	DeleteLiveRecordVodConfigAction   = "DeleteLiveRecordVodConfig"
+	DescribeLiveRecordVodConfigAction = "DescribeLiveRecordVodConfig"
+
 	//直播中心服务器域名
 	DefaultVideoCenter = "video-center.alivecdn.com"
 )
@@ -157,38 +162,24 @@ func (l *Live) cloneRequest(action string) (req *Request) {
 	return
 }
 
-// @see StreamsPublishListWithApp
-func (l *Live) StreamsPublishList(startTime, endTime time.Time, resp interface{}) (err error) {
-	err = l.StreamsPublishListWithApp(l.liveReq.AppName, startTime, endTime, resp)
-	return
-}
-
-// StreamsPublishListWithApp 获取推流列表
+// StreamsPublishList 获取推流列表
 // @appname 应用名 为空时，忽略此参数
 // @startTime 开始时间
 // @endTime   结束时间
 // @link https://help.aliyun.com/document_detail/27191.html?spm=0.0.0.0.Dm58D2
-func (l *Live) StreamsPublishListWithApp(appname string, startTime, endTime time.Time, resp interface{}) (err error) {
+func (l *Live) StreamsPublishList(startTime, endTime time.Time, resp interface{}) (err error) {
 	req := l.cloneRequest(DescribeLiveStreamsPublishListAction)
-	req.AppName = appname
 	req.SetArgs("StartTime", util.GetISO8601TimeStamp(startTime))
 	req.SetArgs("EndTime", util.GetISO8601TimeStamp(endTime))
 	err = l.rpc.Query(req, resp)
 	return
 }
 
-// @see StreamsOnlineListWithApp
-func (l *Live) StreamsOnlineList(resp interface{}) (err error) {
-	err = l.StreamsOnlineListWithApp(l.liveReq.AppName, resp)
-	return
-}
-
-// StreamsOnlineListWithApp 获取在线流
+// StreamsOnlineList 获取在线流
 // @appname 应用名 为空时，忽略此参数
 // @link  https://help.aliyun.com/document_detail/27192.html?spm=0.0.0.0.7uWhjM
-func (l *Live) StreamsOnlineListWithApp(appname string, resp interface{}) (err error) {
+func (l *Live) StreamsOnlineList(resp interface{}) (err error) {
 	req := l.cloneRequest(DescribeLiveStreamsOnlineListAction)
-	req.AppName = appname
 	err = l.rpc.Query(req, resp)
 	return
 }
@@ -202,36 +193,22 @@ func (l *Live) StreamsBlockList(resp interface{}) (err error) {
 	return
 }
 
-// @see StreamsControlHistoryWithApp
-func (l *Live) StreamsControlHistory(startTime, endTime time.Time, resp interface{}) (err error) {
-	err = l.StreamsControlHistoryWithApp(l.liveReq.AppName, startTime, endTime, resp)
-	return
-}
-
-// StreamsControlHistoryWithApp 获取控制历史
+// StreamsControlHistory 获取控制历史
 // @appname 应用名 为空时，忽略此参数
 // @link  https://help.aliyun.com/document_detail/27194.html?spm=0.0.0.0.4DUTT7
-func (l *Live) StreamsControlHistoryWithApp(appname string, startTime, endTime time.Time, resp interface{}) (err error) {
+func (l *Live) StreamsControlHistory(startTime, endTime time.Time, resp interface{}) (err error) {
 	req := l.cloneRequest(DescribeLiveStreamsControlHistoryAction)
-	req.AppName = appname
 	req.SetArgs("StartTime", util.GetISO8601TimeStamp(startTime))
 	req.SetArgs("EndTime", util.GetISO8601TimeStamp(endTime))
 	err = l.rpc.Query(req, resp)
 	return
 }
 
-// @see StreamOnlineUserNumWithApp
-func (l *Live) StreamOnlineUserNum(streamName string, resp interface{}) (err error) {
-	err = l.StreamOnlineUserNumWithApp(l.liveReq.AppName, streamName, resp)
-	return
-}
-
-// StreamOnlineUserNumWithApp 获取在线人数
+// StreamOnlineUserNum 获取在线人数
 // @appname 应用名 为空时，忽略此参数
 // @link https://help.aliyun.com/document_detail/27195.html?spm=0.0.0.0.n6eAJJ
-func (l *Live) StreamOnlineUserNumWithApp(appname string, streamName string, resp interface{}) (err error) {
+func (l *Live) StreamOnlineUserNum(streamName string, resp interface{}) (err error) {
 	req := l.cloneRequest(DescribeLiveStreamOnlineUserNumAction)
-	req.AppName = appname
 	if "" != streamName {
 		req.SetArgs("StreamName", streamName)
 	}
@@ -263,11 +240,6 @@ func (l *Live) ForbidLiveStreamWithPublisher(streamName string, resumeTime *time
 	return l.ForbidLiveStream(l.liveReq.AppName, streamName, "publisher", resumeTime, resp)
 }
 
-// @see ForbidLiveStream
-func (l *Live) ForbidLiveStreamWithPublisherWithApp(appName, streamName string, resumeTime *time.Time, resp interface{}) (err error) {
-	return l.ForbidLiveStream(appName, streamName, "publisher", resumeTime, resp)
-}
-
 // ResumeLiveStream 恢复流
 func (l *Live) ResumeLiveStream(appName, streamName string, liveStreamType string, resp interface{}) (err error) {
 	if global.EmptyString == appName {
@@ -284,11 +256,6 @@ func (l *Live) ResumeLiveStream(appName, streamName string, liveStreamType strin
 // @see ResumeLiveStream
 func (l *Live) ResumeLiveStreamWithPublisher(streamName string, resp interface{}) (err error) {
 	return l.ResumeLiveStream(l.liveReq.AppName, streamName, "publisher", resp)
-}
-
-// @see ResumeLiveStream
-func (l *Live) ResumeLiveStreamWithPublisherWithApp(appName, streamName string, resp interface{}) (err error) {
-	return l.ResumeLiveStream(appName, streamName, "publisher", resp)
 }
 
 // GET 和 SET
