+++ device/live/livepull.go
@@ -10,18 +10,9 @@ import (
 
 // AddLivePullStreamInfoConfig 添加拉流信息
 //
-// @see AddLivePullStreamInfoConfigWithApp
-func (l *Live) AddLivePullStreamInfoConfig(streamName, sourceUrl string, startTime, endTime time.Time, resp interface{}) (err error) {
-	err = l.AddLivePullStreamInfoConfigWithApp(l.liveReq.AppName, streamName, sourceUrl, startTime, endTime, resp)
-	return
-}
-
-// AddLivePullStreamInfoConfigWithApp 添加拉流信息。
-//
 // https://help.aliyun.com/document_detail/57734.html?spm=5176.doc57733.6.656.YS8uOK
-func (l *Live) AddLivePullStreamInfoConfigWithApp(appName, streamName, sourceUrl string, startTime, endTime time.Time, resp interface{}) (err error) {
+func (l *Live) AddLivePullStreamInfoConfig(streamName, sourceUrl string, startTime, endTime time.Time, resp interface{}) (err error) {
 	req := l.cloneRequest(AddLivePullStreamInfoConfigAction)
-	req.AppName = appName
 	req.SetArgs("StreamName", streamName)
 	req.SetArgs("SourceUrl", sourceUrl)
 	req.SetArgs("StartTime", util.GetISO8601TimeStamp(startTime))
@@ -32,18 +23,9 @@ func (l *Live) AddLivePullStreamInfoConfigWithApp(appName, streamName, sourceUrl
 
 // DeleteLivePullStreamInfoConfig 删除拉流信息
 //
-// @see DeleteLivePullStreamInfoConfigWithApp
-func (l *Live) DeleteLivePullStreamInfoConfig(streamName string, resp interface{}) (err error) {
-	err = l.DeleteLivePullStreamInfoConfigWithApp(l.liveReq.AppName, streamName, resp)
-	return
-}
-
-// DeleteLivePullStreamInfoConfigWithApp 删除拉流信息
-//
 // https://help.aliyun.com/document_detail/57735.html?spm=5176.doc57734.6.657.wRW6P7
-func (l *Live) DeleteLivePullStreamInfoConfigWithApp(appName, streamName string, resp interface{}) (err error) {
+func (l *Live) DeleteLivePullStreamInfoConfig(streamName string, resp interface{}) (err error) {
 	req := l.cloneRequest(DeleteLivePullStreamInfoConfigAction)
-	req.AppName = appName
 	req.SetArgs("StreamName", streamName)
 	err = l.rpc.Query(req, resp)
 	return
